+++ routes/product.routes.js
@@ -1,11 +1,19 @@
 const express = require("express");
 const router = express.Router();
+const multer = require("multer");
 
 const productController = require("../controllers/product.controller");
 
+const storage = multer.diskStorage({
+    destination: (request, file, cb) => cb(null, "uploads/"),
+    filename: (request, file, cb) => cb(null, Date.now() + "_" + file.originalname)
+})
+
+const upload = multer({ storage });
+
 // POST: Register a new product
 router.get("/add", productController.getRegisterProduct);
-router.post("/add", productController.postRegisterProduct);
+router.post("/add", upload.single("imagen"), productController.postRegisterProduct);
 
 // GET: Search products
 router.get('/search', productController.searchP);